requirements.txt
langchain>=1.0.0
langgraph>=0.6.0
langchain-google-genai>=2.1.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
